Extract session workbench without behavior changes
