+++ apps/desktop/src/lib/workbench/SessionWorkbench.svelte
